requirements.txt
# flask related libs
Flask
Flask-PyMongo 
python -m pip install "pymongo[srv]"==3.11 